requirements.txt
-i https://pypi.org/simple
langchain==0.3.17; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.16; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.33; python_version >= '3.9' and python_version < '4.0'
langchain-deepseek==0.0.3; python_version >= '3.9' and python_version < '4.0'
langchain-experimental==0.3.4; python_version >= '3.9' and python_version < '4.0'
langchain-openai==0.3.3; python_version >= '3.9' and python_version < '4.0'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
streamlit==1.41.1; python_version >= '3.9' and python_full_version != '3.9.7'
